visual_search.py
"""
Visual Search Engine
Finds visually similar products using CLIP embeddings and FAISS
"""

from PIL import Image
from typing import List
import numpy as np
from loguru import logger

from app.domain.entities.product import Product as ProductEntity
from app.core.config import settings


class VisualSearchEngine:
    """Engine for finding visually similar products"""
    
    def __init__(self):
        """Initialize the visual search engine"""
        # For MVP, we'll use a placeholder
        # In production, this would initialize CLIP model and FAISS index
        logger.info("Initialized VisualSearchEngine (MVP placeholder)")
        self._index_initialized = False
    
    async def find_similar(
        self, 
        query_image: Image.Image, 
        category: str,
        limit: int = 20
    ) -> List[ProductEntity]:
        """
        Find similar products to the query image
        
        Args:
            query_image: PIL Image to search for
            category: Product category to filter by
            limit: Maximum number of results
            
        Returns:
            List of similar Product entities
        """
        # MVP: Return empty list (will be populated by e-commerce search)
        # In production, this would:
        # 1. Generate embedding from query image using CLIP
        # 2. Search FAISS index for similar embeddings
        # 3. Return top-k products
        
        logger.info(f"Visual search for category: {category}, limit: {limit}")
        return []
    
    async def index_product(self, product_id: str, image: Image.Image, embedding: np.ndarray = None):
        """
        Index a product image for similarity search
        
        Args:
            product_id: Product identifier
            image: Product image
            embedding: Optional pre-computed embedding
        """
        # MVP: No-op
        # In production, this would:
        # 1. Generate embedding if not provided
        # 2. Add to FAISS index
        logger.debug(f"Indexing product {product_id} (MVP placeholder)")

