requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Database
pgvector==0.2.4

# Machine Learning
torch==2.2.0
torchvision==0.17.0
transformers==4.37.0
opencv-python==4.9.0
pillow==10.2.0
numpy==1.26.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Caching
redis==5.0.1
hiredis==2.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
